This is so irritating. Just keep it all in one file for now - we only have one entrypoint in the UI.
