--- code/modules/tailwind/src/generateConfig.ts
@@ -1,6 +1,0 @@
-import { type Config } from 'tailwindcss';
-import { applyThemeToTailwindConfig } from './plugins/TailwindConfigAdapter';
-
-export function generateConfigForContent(content: string[]): Config {
-  return applyThemeToTailwindConfig({ content });
-}
