--- code/modules/tailwind/src/main.ts
@@ -1,3 +1,0 @@
-export { generateConfigForContent } from './generateConfig';
-export { applyThemeToTailwindConfig } from './plugins/TailwindConfigAdapter';
-export { TailwindDuotonePlugin } from './plugins/TailwindDuotonePlugin';
