--- code/modules/tailwind/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": "../../../tsconfig.base.json",
-  "include": ["src/plugins"],
-  "compilerOptions": {
-    "baseUrl": "src",
-    "composite": true,
-    "skipLibCheck": true,
-    "module": "ESNext",
-    "moduleResolution": "bundler",
-    "allowSyntheticDefaultImports": true,
-    "declaration": true,
-    "emitDeclarationOnly": true
-  }
-}
